Fix: Don't broadcast new participant User Nav update
With the way TurboStreamActions works (now), just adding a turbo stream
response will automatically do so both as a stream reponse as well as
as a broadcast. But there's no need to broadcast the User Nav update to
other users. Instead, we just need to update the current user's User
Nav, which is exactly what the "response" bucket on TurboStreamActions
is about.

This was just an oversight before... probably missed around the time we
added the 2 distinct buckets to TurboStreamActions.