Tidy: Improve name of "on_game_start_policy" (from "block")
Today I'm preferring "policy" over "block" because I had a brain fart
and was like "what's a block?!". The word just looked odd haha. So, as I
clarified this for myself, I started thinking how really it's a policy
object. It need not always be a block. It just needs to be any object
that responds to `call`. So calling it a "policy" is better.