Fix "Reload" link on Fleet Mustering notification
Previously, if one happened to have just navigated away from the War
Room at the same time as a new game was started, they'd get the Fleet
muster! notification with a "Reload" link to join back in. But, since
they're now on a different page besides the home page, the reload
operation would just reload the current page, instead of showing the
War Room. This is easily fixed by just always navigating to the home
page instead of simply refreshing the current page.

As this was the only place the page#reload action was being called, I've
also removed this action.