Fix: Don't cache just-ended Game "New Game Actions" content
We had previously added caching to make it so that the Just-ended game
message was consistent for all players and across all page refreshes in
53e0fcec44f95b9c1d1224003d733c5003b5c040.
e.g. if the game-ending message was `Argh! 😧 Whhhyyyyyy?!` then it should be
that for everyone and even if they refresh.

However, this was too broad of a fragment cache, as it also included the
"New Game Actions" content. Which includes the ability to create a
"Custom" game. Which is supposed to be reserved only for signers. So, if
a signer and a non-signer were playing together, then whoever didn't end
the Game would see the others "New Game Actions" content, due to the
over-zealous fragment cache.

To fix, we reduce the scope of the overall fragment caching done in
app/views/games/just_ended/_content.html.erb.

Then, in order to preserve the spirit of the change from
53e0fcec44f95b9c1d1224003d733c5003b5c040, we add a second, much more
narrowly scoped fragment cache that just spans the Game Over message.

This way, we still get the benefit of having cached the Game Status
and Game Container partials for the benefit of all observers.

Further, to prevent this from accidentally happening again in the
future, we add `uncacheable!` to the New Game Actions partial
(app/views/games/new/_content.html.erb).
- This sounds like a new best practice I should start employing.