Don't wrap New Game buttons on `md` breakpoint
Previously, we were wrapping the New Game buttons on xs-md breakpoints.
This was because, just before the sm/md breakpoint jump, the buttons
touch the edges of the screen.

However, wrapping for `md` breakpoint looks dumb on landsacpe mobile,
which is the primary usecase for the "md" breakpoint. So, we'll just
ignore the awkwardness on desktop browsers and go with what looks best
in the primary use case (mobile).