Update to Ruby v3.4.6 + bump gems
Just keeping up with the latest so we never fall behind.

Of note, to get Kamal to be able to deploy the app still, had to modify
Dockerfile to COPY in `.ruby-version` before running bundle install.
This is because we use `ruby file: ".ruby-version"` in the Gemfile.

Also, this bundle update includes my latest styles from my gemwork gem.
So I've update many offenses to keep up to date with my latest style
guide.